test: mv shared functions to utils; split compare to own test file
